companies errors handling done or almost

+++ routes/companies.js
@@ -3,8 +3,11 @@ const router = express.Router();
 const models = require("../models");
 
 router.get('/', (req, res) => {
-	models.companies.findAll().then(data => res.json(data))
-})
+	models.companies.findAll()
+	.then(data => {
+		res.status(200).json(data)
+	});
+});
 
 router.get('/:id(\\d+)', (req, res) => {
 	models.companies.findAll({
@@ -12,7 +15,9 @@ router.get('/:id(\\d+)', (req, res) => {
 			id : req.params.id
 		}
 	})
-	.then(data => res.json(data));
+	.then(data => {
+		res.status(200).json(data)
+	});
 })
 
 router.get('/:name', (req, res) => {
@@ -21,7 +26,9 @@ router.get('/:name', (req, res) => {
 			name : req.params.name
 		}
 	})
-	.then(data => res.json(data));
+	.then(data => {
+		res.status(200).json(data)
+	});
 })
 
 router.post('/', (req, res) => {
@@ -39,22 +46,42 @@ router.post('/', (req, res) => {
 })
 
 router.put('/:id(\\d+)', (req, res) => {
-	const data = req.body;
-	console.log(data);
-	models.companies.update(
-		data,
-		{ where : { id : req.params.id } }
-	)
-})
-
-router.delete('/:id(\\d+)', (req, res) => {
-	models.companies.destroy({
-		where : {
-			id : req.params.id
+	models.companies.findById(req.params.id)
+	.then(companiesFound => {
+		if(companiesFound){
+			const data = req.body;
+			console.log(data);
+			models.companies.update(
+				data,
+				{ where : { id : req.params.id } }
+			)
+			.then(updatedCompany => {
+				res.status(200).send(`Company updated at id : ${req.params.id }`);
+			});
+		}
+		else{
+			return res.status(404).send('Company does not exist in DB');
 		}
 	});
+});
 
-	res.sendStatus(200);
-})
+router.delete('/:id(\\d+)', (req, res) => {
+	models.companies.findById(req.params.id)
+	.then(companiesFound => {
+		if(companiesFound){
+			models.companies.destroy({
+				where : {
+					id : req.params.id
+				}
+			})
+			.then(updatedCompany => {
+				res.status(200).send(`Company deleted at id : ${req.params.id }`);
+			})
+		}
+		else{
+			return res.status(404).send('Company does not exist in DB');
+		}
+	});
+});
 
 module.exports = router;
